Split Property provider into controller/provider
